+++ 1.21.5/Common/src/main/java/fuzs/puzzleslib/api/client/gui/v2/tooltip/ItemTooltipRegistry.java
@@ -128,6 +128,8 @@ public final class ItemTooltipRegistry {
 
         /**
          * A tooltip provider for adding individual tooltip lines.
+         * <p>
+         * TODO add player
          *
          * @param itemLike            the item / block
          * @param itemStack           the item stack