+++ 1.21.5/Common/src/main/java/fuzs/puzzleslib/api/event/v1/core/EventPhase.java
@@ -5,7 +5,6 @@ import fuzs.puzzleslib.impl.PuzzlesLibMod;
 import fuzs.puzzleslib.impl.event.core.EventPhaseImpl;
 import net.minecraft.resources.ResourceLocation;
 import org.jetbrains.annotations.ApiStatus;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.function.BiConsumer;
 
@@ -18,7 +17,9 @@ public interface EventPhase {
     /**
      * Fabric's default event phase, equivalent to EventPriority#NORMAL on Forge.
      */
-    EventPhase DEFAULT = new EventPhaseImpl(ResourceLocationHelper.fromNamespaceAndPath("fabric", "default"), null, null);
+    EventPhase DEFAULT = new EventPhaseImpl(ResourceLocationHelper.fromNamespaceAndPath("fabric", "default"),
+            null,
+            null);
     /**
      * A phase to be used directly before the default phase, equivalent to EventPriority#HIGH on Forge.
      */
@@ -45,15 +46,26 @@ public interface EventPhase {
      * @return another event phase that runs before / after this one, the order is defined by
      *         {@link #applyOrdering(BiConsumer)}
      */
-    @Nullable EventPhase parent();
+    EventPhase parent();
 
     /**
      * The ordering defines in which relation this event phase is to {@link #parent()}, if it is supposed to run before
      * or afterward.
      *
-     * @param consumer apply event phases to the Fabric event
+     * @param phaseOrderingConsumer apply event phases to the Fabric event
      */
-    void applyOrdering(BiConsumer<ResourceLocation, ResourceLocation> consumer);
+    default void applyOrdering(BiConsumer<ResourceLocation, ResourceLocation> phaseOrderingConsumer) {
+        this.applyOrdering(this.resourceLocation(), phaseOrderingConsumer);
+    }
+
+    /**
+     * The ordering defines in which relation this event phase is to {@link #parent()}, if it is supposed to run before
+     * or afterward.
+     *
+     * @param resourceLocation      the resource location used for registering this phase on Fabric
+     * @param phaseOrderingConsumer apply event phases to the Fabric event
+     */
+    void applyOrdering(ResourceLocation resourceLocation, BiConsumer<ResourceLocation, ResourceLocation> phaseOrderingConsumer);
 
     /**
      * @return the ordering value used in relation to {@link #parent()}
@@ -67,8 +79,9 @@ public interface EventPhase {
      * @return the custom event phase
      */
     static EventPhase early(EventPhase eventPhase) {
-        return new EventPhaseImpl(PuzzlesLibMod.id(
-                "early_" + eventPhase.resourceLocation().getPath()), eventPhase, EventPhaseImpl.Ordering.BEFORE);
+        return new EventPhaseImpl(PuzzlesLibMod.id("early_" + eventPhase.resourceLocation().getPath()),
+                eventPhase,
+                EventPhaseImpl.Ordering.BEFORE);
     }
 
     /**
@@ -78,7 +91,8 @@ public interface EventPhase {
      * @return the custom event phase
      */
     static EventPhase late(EventPhase eventPhase) {
-        return new EventPhaseImpl(PuzzlesLibMod.id(
-                "late_" + eventPhase.resourceLocation().getPath()), eventPhase, EventPhaseImpl.Ordering.AFTER);
+        return new EventPhaseImpl(PuzzlesLibMod.id("late_" + eventPhase.resourceLocation().getPath()),
+                eventPhase,
+                EventPhaseImpl.Ordering.AFTER);
     }
 }