+++ 1.21.5/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.5.8-1.21.5] - 2025-04-15
+### Added
+- Add `ContainerMenuHelper::openMenu`
+### Fixed
+- Only trigger `CommandOverrides` when commands are allowed
+- Fix event phase converters not respecting the ordering of all parent event phases
+
 ## [v21.5.7-1.21.5] - 2025-04-14
 ### Changed
 - Rework implementation for menu types with additional data sent to clients